requirements.txt
astroid             >=2.2.5
certifi             >=2019.6.16
chardet             >=3.0.4
colorama            >=0.4.1
defusedxml          >=0.6.0
Django              >=2.2.3
django-allauth      >=0.39.1
django-cors-headers >=3.0.2
django-crispy-forms >=1.7.2
django-rest-auth    >=0.9.5
djangorestframework >=3.10.1
idna                >=2.8
isort               >=4.3.21
lazy-object-proxy   >=1.4.1
mccabe              >=0.6.1
oauthlib            >=3.0.2
pip                 >=19.0.3
PyJWT               >=1.7.1
pylint              >=2.3.1
PySocks             >=1.7.0
python3-openid      >=3.1.0
pytz                >=2019.1
requests            >=2.22.0
requests-oauthlib   >=1.2.0
setuptools          >=40.8.0
six                 >=1.12.0
sqlparse            >=0.3.0
twilio              >=6.29.1
typed-ast           >=1.4.0
urllib3             >=1.25.3
wrapt               >=1.11.2
